[desktop] Add localizedValue method. JB#56277 OMP#JOLLA-529
This allows fetching any key with a localized value.

Instead of using a specific for Name translation, add a template that
can be used for any key's translation id. Also remove some MeeGo legacy
by introducing Amber name for translation catalog.

As translator can be now used multiple times but it's unlikely to be
used after a longer period of time, thus a pointer to it is stored and
the instance is destroyed with a timer after a while. If translations
are needed later again, the translator is constructed again.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>