[mlite] Rename MGconfItem as MDConfItem. JB#56217
The backend was changed to DConf in 2014. About time we provide a
properly named class here. Renaming as first step to keep git history
on the file. On follow-up commit reintroducing the old class back
for backward compatibility.

Downside with this setup is duplicate content for the classes but
anything fancier would have needed care not to break ABI or
resulting other complication. The class itself is quite small
and rarely getting changes so think this should be good enough.